Split `AlertModals` into multiple files (#46376)
