+++ frontend/src/metabase/query_builder/components/AlertModals/AlertAboveGoalToggle.jsx
@@ -1,0 +1,5 @@
+import { AlertSettingToggle } from "./AlertSettingToggle";
+
+export const AlertAboveGoalToggle = props => (
+  <AlertSettingToggle {...props} setting="alert_above_goal" />
+);
