+++ frontend/src/metabase/query_builder/components/AlertModals/AlertFirstOnlyToggle.jsx
@@ -1,0 +1,5 @@
+import { AlertSettingToggle } from "./AlertSettingToggle";
+
+export const AlertFirstOnlyToggle = props => (
+  <AlertSettingToggle {...props} setting="alert_first_only" />
+);
