+++ frontend/src/metabase/query_builder/components/AlertModals/AlertModals.module.css
@@ -2,3 +2,7 @@
   border-width: 2px;
   border-color: var(--mb-color-border);
 }
+
+.AlertModalsTextWidth {
+  width: 162px;
+}
