+++ frontend/src/metabase/query_builder/components/AlertModals/index.ts
@@ -1,1 +1,15 @@
-export * from "./AlertModals";
+export { CreateAlertModalContent } from "./CreateAlertModalContent";
+export { UpdateAlertModalContent } from "./UpdateAlertModalContent";
+export { AlertEditForm } from "./AlertEditForm";
+export { AlertModalTitle } from "./AlertModalTitle";
+export { AlertAboveGoalToggle } from "./AlertAboveGoalToggle";
+export { AlertFirstOnlyToggle } from "./AlertFirstOnlyToggle";
+export { AlertSettingToggle } from "./AlertSettingToggle";
+export { AlertEditChannels } from "./AlertEditChannels";
+export { NormalAlertTip } from "./NormalAlertTip";
+export { MultiSeriesAlertTip } from "./MultiSeriesAlertTip";
+export { RawDataAlertTip } from "./RawDataAlertTip";
+export { AlertGoalToggles } from "./AlertGoalToggles";
+export { DeleteAlertSection } from "./DeleteAlertSection";
+export { AlertEducationalScreen } from "./AlertEducationalScreen";
+export { AlertEditSchedule } from "./AlertEditSchedule";
